BID/requirement.txt
django == 1.8.3
django-compressor == 1.5
djangorestframework == 3.1.3
django-filter == 0.10.0
Markdown == 2.6.2
psycopg2 == 2.7.4
#tornado == 4.2
#redis == 2.10.3
#tornado-redis ==2.4.18
requests